Refactor chat app state into contexts
